requirements.txt
# Core libraries
fastapi==0.109.0
uvicorn==0.24.0
python-socketio==5.10.0
python-dotenv==1.0.0

# Semantic Kernel
semantic-kernel==0.4.7.dev0

# OpenAI
openai==1.12.0

# Tracing and monitoring
azure-monitor-opentelemetry==1.1.1
opentelemetry-instrumentation-openai==0.1.4

# Other utilities
python-multipart==0.0.6
